Update API for simplicity
