+++ API/InsiteTeamTask.Data/Models/Game.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace InsiteTeamTask.Models
+﻿namespace InsiteTeamTask.Data.Models
 {
     public class Game
     {
