+++ API/InsiteTeamTask.Data/Models/Season.cs
@@ -1,0 +1,7 @@
+﻿namespace InsiteTeamTask.Data.Models
+{
+    public class Season
+    {
+        public int Id { get; set; }
+    }
+}
