+++ API/InsiteTeamTask.Data/Models/Ticket.cs
@@ -1,0 +1,9 @@
+﻿namespace InsiteTeamTask.Data.Models
+{
+    public class Ticket
+    {
+        public string Barcode { get; set; }
+
+        public string ProductId { get; set; }
+    }
+}
