+++ API/InsiteTeamTask.Tests/InsiteTeamTask.Tests.csproj
@@ -13,7 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\InsiteTeamTask\InsiteTeamTask.csproj" />
+    <ProjectReference Include="..\InsiteTeamTask\InsiteTeamTask.API.csproj" />
   </ItemGroup>
 
 </Project>
