+++ API/InsiteTeamTask.sln
@@ -3,9 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28307.539
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InsiteTeamTask", "InsiteTeamTask\InsiteTeamTask.csproj", "{21ADB46E-89E9-4424-9C1D-78BC0968BA37}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InsiteTeamTask.API", "InsiteTeamTask\InsiteTeamTask.API.csproj", "{21ADB46E-89E9-4424-9C1D-78BC0968BA37}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InsiteTeamTask.Tests", "InsiteTeamTask.Tests\InsiteTeamTask.Tests.csproj", "{761B9D68-0962-4E2C-9C95-DBFE6F897CDB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InsiteTeamTask.Tests", "InsiteTeamTask.Tests\InsiteTeamTask.Tests.csproj", "{761B9D68-0962-4E2C-9C95-DBFE6F897CDB}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InsiteTeamTask.Data", "InsiteTeamTask.Data\InsiteTeamTask.Data.csproj", "{CDDA3760-B4B1-41FC-8302-57033F99F4DA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{761B9D68-0962-4E2C-9C95-DBFE6F897CDB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{761B9D68-0962-4E2C-9C95-DBFE6F897CDB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{761B9D68-0962-4E2C-9C95-DBFE6F897CDB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CDDA3760-B4B1-41FC-8302-57033F99F4DA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CDDA3760-B4B1-41FC-8302-57033F99F4DA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CDDA3760-B4B1-41FC-8302-57033F99F4DA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CDDA3760-B4B1-41FC-8302-57033F99F4DA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
