+++ API/InsiteTeamTask/InsiteTeamTask.API.csproj
@@ -10,4 +10,8 @@
     <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\InsiteTeamTask.Data\InsiteTeamTask.Data.csproj" />
+  </ItemGroup>
+
 </Project>
