+++ API/InsiteTeamTask/Services/IDataService.cs
@@ -1,0 +1,14 @@
+﻿using InsiteTeamTask.Models;
+using System.Collections.Generic;
+
+namespace InsiteTeamTask.Services
+{
+    public interface IDataService
+    {
+        List<Attendance> GetAttendance();
+
+        List<Attendance> GetAttendanceForGame(int seasonId, int gameNumber);
+
+        List<Attendance> GetAttendanceForProduct(string productCode);
+    }
+}
