--- InsiteTeamTask.Tests/InsiteTeamTaskTests.cs
@@ -1,26 +1,0 @@
-using InsiteTeamTask.Repositories;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System.Linq;
-
-namespace InsiteTeamTask.Tests
-{
-    [TestClass]
-    public class InsiteTeamTaskTests
-    {
-        [TestMethod]
-        public void TestMethod1()
-        {
-            // Arrange
-            var mockDataService = new MockDataService();
-            var repo = new DataRepository();
-            var games = mockDataService.Games();
-
- 
-            // Act
-            var attendanceList = repo.GetAttendanceListFor(games.First().Id);
-
-            // Assert
-            Assert.IsNotNull(attendanceList);
-        }
-    }
-}
