--- InsiteTeamTask/Models/Event.cs
@@ -1,6 +1,0 @@
-﻿namespace InsiteTeamTask.Models
-{
-    public class Event
-    {
-    }
-}
