--- InsiteTeamTask/Models/Season.cs
@@ -1,11 +1,0 @@
-﻿namespace InsiteTeamTask.Models
-{
-    public class Season
-    {
-        public Season(int id)
-        {
-            this.Id = id;
-        }
-        public int Id { get; set; }
-    }
-}
