--- InsiteTeamTask/Models/Ticket.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace InsiteTeamTask.Models
-{
-    public class Ticket
-    {
-        public string Barcode { get; set; }
-
-        public string ProductId { get; set; }
-
-    }
-}
